+++ src/main/java/org/geowe/service/geometry/JtsDivideService.java
@@ -15,6 +15,7 @@
  ******************************************************************************/
 package org.geowe.service.geometry;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.validation.Valid;
@@ -31,6 +32,7 @@ import org.geowe.service.filter.DivisionFilter;
 import org.geowe.service.geometry.engine.GeoEngineer;
 import org.geowe.service.geometry.engine.JTSGeoEngineer;
 import org.geowe.service.model.DivisionData;
+import org.geowe.service.model.FlatGeometry;
 import org.jboss.resteasy.annotations.GZIP;
 
 /**
@@ -43,6 +45,20 @@ import org.jboss.resteasy.annotations.GZIP;
 public class JtsDivideService {
 
 	
+	@POST
+	@GZIP
+	@Produces(MediaType.APPLICATION_JSON)
+	@Consumes(MediaType.APPLICATION_JSON)
+	public Response divide(@NotNull @Valid FlatGeometry flatGeometry) {
+		GeoEngineer geoEngineer = new JTSGeoEngineer();
+		List<String> dividedGeometries = geoEngineer.decompose(flatGeometry.getWkt());
+
+		return Response.status(Status.CREATED)
+				.entity(getFlatGeometries(dividedGeometries, flatGeometry))
+				.build();
+	}
+	
+	
 	@Path("/polygon")
 	@POST
 	@GZIP
@@ -52,13 +68,12 @@ public class JtsDivideService {
 	public Response dividePolygons(@NotNull @Valid DivisionData divisionData) {
 		GeoEngineer geoEngineer = new JTSGeoEngineer();
 		List<String> dividedPolygons = geoEngineer.dividePolygon(divisionData);
-		
+
 		return Response.status(Status.CREATED)
-				.entity(dividedPolygons)
+				.entity(getFlatGeometries(dividedPolygons, divisionData.getGeomToDivide()))
 				.build();
 	}
-	
-	
+
 	@Path("/line")
 	@POST
 	@GZIP
@@ -69,9 +84,16 @@ public class JtsDivideService {
 		GeoEngineer geoEngineer = new JTSGeoEngineer();
 		List<String> dividedLines = geoEngineer.divideLine(divisionData);
 		return Response.status(Status.CREATED)
-				.entity(dividedLines)
+				.entity(getFlatGeometries(dividedLines, divisionData.getGeomToDivide()))
 				.build();
 	}
-	
-}
 
+	private List<FlatGeometry> getFlatGeometries(List<String> dividedWkts, FlatGeometry flatGeometry) {
+		List<FlatGeometry> flatGeometries = new ArrayList<FlatGeometry>();
+		dividedWkts.forEach(wkt -> flatGeometries
+				.add(new FlatGeometryBuilder().id(flatGeometry.getId())
+						.crs(flatGeometry.getCrs()).wkt(wkt).build()));
+		return flatGeometries;
+	}
+
+}