+++ CHANGELOG.md
@@ -9,6 +9,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Geometry validations
 - Asynchronous tasks
 
+## [0.7.0]SNAPSHOT - 2017-03-17
+
+### Added
+- Implemented resources for JTS operations:
+	- Decompose multi geometry into basic geometry /jts/division
+
+- Divide LineString resource (/jts/division/line) return FlatGeometries
+- Divide Polygon (/jts/division/polygon) return FlatGeometries
+	
+	
 ## [0.6.0]SNAPSHOT - 2017-03-16
 
 ### Added